Squash sqeleton's utils & data_diff's utils
