+++ data_diff/sqeleton/databases/oracle.py
@@ -1,6 +1,6 @@
 from typing import Dict, List, Optional
 
-from data_diff.sqeleton.utils import match_regexps
+from data_diff.utils import match_regexps
 from data_diff.sqeleton.abcs.database_types import (
     Decimal,
     Float,
