+++ data_diff/table_segment.py
@@ -7,7 +7,7 @@ from runtype import dataclass
 from typing_extensions import Self
 
 from data_diff.utils import safezip, Vector
-from data_diff.sqeleton.utils import ArithString, split_space
+from data_diff.utils import ArithString, split_space
 from data_diff.sqeleton.databases import Database, DbPath, DbKey, DbTime
 from data_diff.sqeleton.schema import Schema, create_schema
 from data_diff.sqeleton.queries import Count, Checksum, SKIP, table, this, Expr, min_, max_, Code
