+++ tests/test_database_types.py
@@ -13,7 +13,7 @@ from itertools import islice, repeat, chain
 
 from parameterized import parameterized
 
-from data_diff.sqeleton.utils import number_to_human
+from data_diff.utils import number_to_human
 from data_diff.sqeleton.queries import table, commit, this, Code
 from data_diff.sqeleton.queries.api import insert_rows_in_batches
