+++ tests/test_diff_tables.py
@@ -4,7 +4,7 @@ import uuid
 import unittest
 
 from data_diff.sqeleton.queries import table, this, commit, code
-from data_diff.sqeleton.utils import ArithAlphanumeric, numberToAlphanum
+from data_diff.utils import ArithAlphanumeric, numberToAlphanum
 
 from data_diff.hashdiff_tables import HashDiffer
 from data_diff.joindiff_tables import JoinDiffer
